Created npm_requirements.txt and improved Makefile

+++ Makefile
@@ -6,12 +6,17 @@ run:
 	. $(VIRTUALENV)bin/activate ; export PYTHONPATH=`pwd`/app/:$(PYTHONPATH) ; python -W ignore::DeprecationWarning manage.py runserver
 
 install:
+	# if venv folder is not created yet we do it
 	if [ ! -d venv ]; then virtualenv venv --no-site-packages --distribute --prompt=Arte-WoW; fi
 
 	# Install pip packages
 	. $(VIRTUALENV)bin/activate; pip install -r requirements.txt
 
+	make npm_install
+
+npm_install:
 	# Install npm packages
-	xargs -a npm_requirements.txt npm install -g
+	if [ -s npm_requirements.txt ]; then xargs -a npm_requirements.txt npm install -g; else echo '\nNo NPM dependencies found in npm_requirements.txt'; fi
+
 
 # EOF
